loading screen refactoring
